First commit
The Top Gun website was created as a project for a module called Scripting for interactive media in the 3rd year at Edinburgh Napier University. The project consisted of using front-end technologies, HTML, JQuery, and CSS. The website was mainly informational in the style of an upcoming movie premiere. The information posted on the website was mainly about the cast, story, and the premiere.